+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Add 502 and 504 status codes as retriable
 ([#1153](https://github.com/census-instrumentation/opencensus-python/pull/1153))
 - Fix statsbeat bug - exporting zero values for network statsbeat
-([#1153](https://github.com/census-instrumentation/opencensus-python/pull/1153))
+([#1155](https://github.com/census-instrumentation/opencensus-python/pull/1155))
 
 ## 1.1.6
 Released 2022-08-03